+++ .env.sample
@@ -4,5 +4,5 @@ TAPYRUS_RPC_PORT=12381
 TAPYRUS_RPC_USER=rpcuser
 TAPYRUS_RPC_PASSWORD=rpcpassword
 TAPYRUS_AUTHORITY_KEY=cUJN5RVzYWFoeY8rUztd47jzXCu1p57Ay8V7pqCzsBD3PEXN7Dd4
-BRAND_URL=http://localhost:3000
+BRAND_URL=http://localhost:3000/api
 DID_SERVICE_URL=http://localhost:4000
